Refactor code as classes
