+++ data/duke.txt
@@ -1,5 +1,6 @@
 T | 1 | read book
-D | 1 |return book | June 6th
-E | 1 |project meeting | Aug 6th 2-4pm
-T | 1 | join sports club
-D | 0 | return book | Oct 15 2019, 18:00
+D | 1 | return book | June 6th
+T | 1 | eat
+T | 0 | sleep
+T | 1 | eat
+T | 0 | sleep
