+++ src/main/java/DukeException.java
@@ -1,0 +1,5 @@
+public class DukeException extends Exception {
+    public DukeException(String e) {
+        super(e);
+    }
+}
